# .agents/skills/payluk-api/references/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Payluk ThirdParty API",
    "version": "1.0.0",
    "description": "Merchant / Business API for the Payluk escrow and payments platform.\n\nAll routes are mounted under `/v1` and authenticated with a secret key. Every response uses the standard envelope `{ status, message, data }`.\n\n**Rate limit.** Each secret key may make **10 requests per minute** across all `/v1` routes. Beyond that the API answers `429` with `{ \"status\": 429, \"message\": \"Too many request\" }`. Every response carries `RateLimit` and `RateLimit-Policy` headers (IETF draft-7) so a client can back off before being cut off.\n\n**IP allowlist.** On production, if you have added IP addresses to the allowlist in your dashboard, requests from any other address are refused with `403 Unauthorized IP address`. An empty allowlist admits every address. Staging never checks the caller's IP.\n\n**Merchant super-admin scope.** Escrow, milestone-escrow and category routes accept any approved business key. Every other route (customers, wallet, merchant account, confirming and disputing, additional fee, vault, payments, cards, countries, crypto whitelist) requires the key to belong to a merchant super-admin account and answers `401 Access denied` otherwise.",
    "contact": {
      "name": "Payluk Developer Support",
      "email": "support@payluk.ng",
      "url": "https://payluk.ng"
    }
  },
  "servers": [
    {
      "url": "https://staging.api.payluk.ng",
...
    },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/escrow/dispute/resolve/{escrowId}": {
      "post": {
        "tags": [
          "Disputes"
        ],
        "summary": "Resolve dispute",
        "operationId": "resolveDispute",
        "description": "Merchant resolves a dispute: `COMPLETED` releases to the seller; `REFUNDED` returns funds to the buyer; `SPLIT` divides the held funds between them in amounts you name. Must **not** send a `customer-id` header; the escrow must belong to one of the merchant's customers. Sent as `multipart/form-data`.\n\n### Splitting the held funds\n\nWhen neither side is wholly right, send `status: SPLIT` with `sellerAmount` and `buyerAmount`. The two must add up to exactly what the escrow still holds, which is **not** always `amount`:\n\n- **Standard escrow**: `amount` minus the seller's share of the escrow fee. On a buyer-paid escrow the seller pays no share, so the pool is the full `amount`.\n- **Milestone escrow**: the total of the milestones the buyer never confirmed. Already-released milestones stay with the seller and are untouched.\n\nRead the pool off the escrow before you rule, and send amounts with at most two decimal places. Amounts that do not add up are rejected and no money moves. Vault escrows cannot be split.\n\nThe escrow closes as `SPLIT`, its still-held milestones are marked `SPLIT`, a `split` object records the division, and an `escrow.split` webhook fires.\n\n### The additional (delivery) fee\n\nIf the escrow carries an [additional fee](/api-reference/escrow/update-additional-fee), **this is where you decide who keeps it**. By the time you rule on a dispute you know whether the delivery actually happened, which you could not know when you priced it.\n\nSend `additionalFeeRefundable` with the resolution:\n\n- **`true`**: the fee goes back to the buyer along with the refunded principal. Nothing was delivered.\n- **`false`**: the fee is credited to your merchant wallet as if the escrow had completed. The delivery cost was already incurred.\n\nYour choice is applied to this settlement **and saved on the escrow**, overriding whatever was set before payment. Omit the field to settle on the escrow's existing `additionalFeeRefundable` (which defaults to `true`). `REFUNDED` and `SPLIT` resolutions are affected; a `COMPLETED` one always pays the fee to you.\n\nThe escrow fee itself is never refunded to either party, on a split any more than on a refund; see [Fees & settlement](/concepts/fees-and-settlement).",
        "parameters": [
          {
            "$ref": "#/components/parameters/EscrowId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "resolution",
                  "status"
                ],
                "properties": {
                  "resolution": {
                    "type": "string",
                    "example": "Refund approved after review"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "COMPLETED",
                      "REFUNDED",
                      "SPLIT"
                    ],
                    "example": "REFUNDED"
                  },
                  "sellerAmount": {
                    "type": "number",
                    "description": "Required on a `SPLIT` resolution, rejected otherwise. The part of the held funds released to the seller. At most two decimal places. Together with `buyerAmount` it must equal exactly what the escrow still holds.",
                    "example": 30000
                  },
                  "buyerAmount": {
                    "type": "number",
                    "description": "Required on a `SPLIT` resolution, rejected otherwise. The part of the held funds returned to the buyer.",
                    "example": 20000
                  },
                  "additionalFeeRefundable": {
                    "type": "boolean",
                    "description": "Optional, and only meaningful on a `REFUNDED` or `SPLIT` resolution. `true` returns the escrow's `additionalFee` to the buyer, `false` credits it to your merchant wallet. Saved on the escrow, overriding the value set before payment. Omit to use whatever the escrow already carries.",
                    "example": true
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional supporting document."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispute resolved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Dispute resolution submitted successfully",
                  "data": {
                    "id": "6948fb43cb570eeef76285e3",
                    "amount": 1000,
                    "purpose": "iPhone 13 Pro Max",
                    "description": "Black Titanium, White Titanium, Natural Titanium, Desert Titanium",
                    "whoPays": "buyer",
                    "imageUrl": [
                      "https://mediacloud.me/media/bWVkaWEvaW1hZ2VzL29wdGltaXplL3Z0VjhwN0psSzlIMzhEVWxicWpmV1JZdUpuSUZNWlhEdU03UkVmeHEuanBn.jpg"
                    ],
                    "fee": 15,
                    "paymentToken": "PY_KrWLqPd90314",
                    "paidAt": "1766391073",
                    "status": "REFUNDED",
                    "state": "CLOSED",
                    "logs": [
                      "DISPUTED"
                    ],
                    "channel": "API",
                    "isSeller": false,
                    "dispute": [
                      {
                        "name": "Ezumah",
                        "profile": null,
                        "message": "Item not as described; the seal was broken on arrival",
                        "proofUrl": "https://mediacloud.me/media/dispute-evidence.jpg",
                        "type": "buyer",
                        "createdAt": "2025-12-22 08:20:00"
                      },
                      {
                        "name": "Customer support",
                        "profile": null,
                        "message": "Refund approved after review of the buyer's evidence.",
                        "proofUrl": "https://mediacloud.me/media/resolution-note.pdf",
                        "type": "customer support",
                        "createdAt": "2025-12-22 09:05:00"
                      }
                    ],
                    "paymentDetails": {
                      "id": "6948fd20cb570eeef76286a4",
                      "amount": 1015,
                      "status": "success",
                      "reference": "9500364766547485",
                      "fee": 15,
                      "transactionType": "escrow",
                      "currency": "NGN",
                      "createdAt": "2025-12-22T08:11:12.605Z"
                    },
                    "category": null,
                    "completedAt": "1766394300",
                    "approvedClaimBy": null,
                    "refundedBy": null,
                    "maxDelivery": 20,
                    "deliveryTimeline": "minutes",
                    "totalQuantity": 1,
                    "settlementType": "STANDARD",
                    "milestones": [],
                    "createdAt": "2025-12-22T08:03:15.124Z",
                    "updatedAt": "2025-12-22T09:05:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "No dispute found or escrow already completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "noDispute": {
                    "summary": "No dispute",
                    "value": {
                      "status": 400,
                      "message": "No dispute found for this escrow",
                      "data": {}
                    }
                  },
                  "completed": {
                    "summary": "Already completed",
                    "value": {
                      "status": 400,
                      "message": "Escrow already completed",
                      "data": {}
                    }
                  },
                  "badRefundFlag": {
                    "summary": "Invalid additionalFeeRefundable",
                    "value": {
                      "status": 400,
                      "message": "additionalFeeRefundable must be true or false",
                      "data": {}
                    }
                  },
                  "splitAmountsMissing": {
                    "summary": "SPLIT without amounts",
                    "value": {
                      "status": 400,
                      "message": "sellerAmount and buyerAmount are required on a SPLIT resolution",
                      "data": {}
                    }
                  },
                  "splitAmountsMismatch": {
                    "summary": "Split does not add up",
                    "value": {
                      "status": 400,
                      "message": "A split must add up to the 50000 still held on this escrow",
                      "data": {}
                    }
                  },
                  "splitAmountsUnexpected": {
                    "summary": "Amounts on a non-split resolution",
                    "value": {
                      "status": 400,
                      "message": "sellerAmount and buyerAmount are only valid on a SPLIT resolution",
                      "data": {}
                    }
                  },
                  "splitNotSupported": {
                    "summary": "Settlement model cannot be split",
                    "value": {
                      "status": 400,
                      "message": "A VAULT escrow cannot be split between the parties",
                      "data": {}
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your secret key as a Bearer token. The key prefix selects the environment: `sk_test_...` (staging) or `sk_live_...` (production); a key on the wrong host is refused with `403 Unauthorized Access`. Each key is limited to 10 requests per minute (`429` beyond that) and, on production, to the IP addresses on your dashboard allowlist when one is configured."
      }
    },
    "parameters": {
...
      "EscrowId": {
        "name": "escrowId",
        "in": "path",
        "required": true,
        "description": "The escrow's unique identifier.",
        "schema": {
          "type": "string"
        },
        "example": "665f1b2c9a1e4d0012ab3c10"
      },
...
    },
    "responses": {
...
      "AccessDenied": {
        "description": "Access denied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "status": 401,
              "message": "Access denied",
              "data": {}
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded: more than 10 requests in one minute from this secret key. Read the `RateLimit` header and retry after the window resets.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "status": 429,
              "message": "Too many request"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Standard error envelope.",
        "properties": {
          "status": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Action not allowed"
          },
          "data": {
            "type": "object",
            "example": {}
          }
        }
      },
...
      "EscrowStatus": {
        "type": "string",
        "description": "High-level escrow status.",
        "enum": [
          "PENDING",
          "ONGOING",
          "COMPLETED",
          "REFUNDED",
          "CLAIMED",
          "DISPUTED",
          "INVESTIGATING",
          "SPLIT"
        ]
      },
      "EscrowState": {
        "type": "string",
        "description": "Lifecycle state of the escrow.",
        "enum": [
          "AWAITING_PAYMENT",
          "OPENED",
          "CLOSED"
        ]
      },
      "SettlementType": {
        "type": "string",
        "enum": [
          "STANDARD",
          "MILESTONE",
          "VAULT"
        ]
      },
      "Milestone": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665f1b2c9a1e4d0012ab3c21"
          },
          "title": {
            "type": "string",
            "example": "Design"
          },
          "description": {
            "type": "string",
            "example": "Figma mockups"
          },
          "amount": {
            "type": "integer",
            "example": 300000
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "example": "2026-07-15"
          },
          "customerId": {
            "type": "string",
            "nullable": true,
            "description": "Optional beneficiary. When set, releasing this milestone credits this customer's main balance instead of the seller's.",
            "example": null
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RELEASED",
              "REFUNDED",
              "SPLIT"
            ],
            "description": "`SPLIT` means this milestone was still held when a dispute was resolved by dividing the funds between both parties.",
            "example": "PENDING"
          },
          "releasedAt": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        }
      },
...
      "VaultParticipant": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "example": "665f1b2c9a1e4d0012ab3c40"
          },
          "amount": {
            "type": "integer",
            "example": 600000
          },
          "status": {
            "type": "string",
            "enum": [
              "STAKED",
              "RELEASED",
              "REFUNDED"
            ],
            "example": "STAKED"
          },
          "stakedAt": {
            "type": "string",
            "description": "Unix timestamp (seconds) the stake was locked.",
            "example": "1784900000"
          },
          "settledAt": {
            "type": "string",
            "nullable": true,
            "description": "Unix timestamp (seconds) the stake was released or refunded.",
            "example": null
          }
        }
      },
...
      "Escrow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665f1b2c9a1e4d0012ab3c10"
          },
          "amount": {
            "type": "integer",
            "example": 150000
          },
          "purpose": {
            "type": "string",
            "example": "MacBook Pro 14\""
          },
          "description": {
            "type": "string",
            "example": "Space grey, sealed"
          },
          "whoPays": {
            "type": "string",
            "enum": [
              "buyer",
              "seller",
              "both"
            ],
            "example": "both"
          },
          "imageUrl": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "https://cdn.payluk.ng/escrow/abc.png"
            ]
          },
          "fee": {
            "type": "integer",
            "example": 3750
          },
          "additionalFee": {
            "type": "integer",
            "description": "The merchant's own charge on top of the amount (delivery and the like). 0 unless set through Update additional fee.",
            "example": 0
          },
          "additionalFeeRefundable": {
            "type": "boolean",
            "description": "Whether a buyer-favoured dispute refund returns the additional fee to the buyer.",
            "example": true
          },
          "paymentToken": {
            "type": "string",
            "example": "PY_8AB12C9D3045"
          },
          "paidAt": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "status": {
            "$ref": "#/components/schemas/EscrowStatus"
          },
          "state": {
            "$ref": "#/components/schemas/EscrowState"
          },
          "channel": {
            "type": "string",
            "example": "API"
          },
          "isSeller": {
            "type": "boolean",
            "example": true
          },
          "dispute": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object"
            },
            "example": null
          },
          "category": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "maxDelivery": {
            "type": "integer",
            "example": 3
          },
          "deliveryTimeline": {
            "type": "string",
            "example": "days"
          },
          "totalQuantity": {
            "type": "integer",
            "example": 1
          },
          "settlementType": {
            "$ref": "#/components/schemas/SettlementType"
          },
          "milestones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Milestone"
            }
          },
          "participants": {
            "type": "array",
            "nullable": true,
            "description": "VAULT escrows only: the customers staking into the pot.",
            "items": {
              "$ref": "#/components/schemas/VaultParticipant"
            }
          },
          "winnerId": {
            "type": "string",
            "nullable": true,
            "description": "VAULT escrows only: the participant declared the winner by the merchant.",
            "example": null
          },
          "split": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EscrowSplit"
              }
            ],
            "nullable": true,
            "description": "Set only when the dispute was resolved with `status: SPLIT`. Null on every other escrow."
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-22T10:15:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-06-22T10:15:00.000Z"
          }
        }
      },
      "EscrowSplit": {
        "type": "object",
        "description": "How a `SPLIT` dispute resolution divided the funds the escrow was still holding.",
        "properties": {
          "sellerAmount": {
            "type": "number",
            "description": "Released to the seller's main balance.",
            "example": 30000
          },
          "buyerAmount": {
            "type": "number",
            "description": "Returned to the buyer's main balance.",
            "example": 20000
          },
          "pool": {
            "type": "number",
            "description": "What the escrow held at the moment of resolution. Always equals `sellerAmount + buyerAmount`.",
            "example": 50000
          },
          "resolvedAt": {
            "type": "string",
            "description": "Unix seconds, as a string.",
            "example": "1786694400"
          }
        }
      },
      "EscrowResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "$ref": "#/components/schemas/Escrow"
          }
        }
      },
...
    }
  }
}